# Upstream memory and cache fixes. Some book loading optimisations