--- a/crengine/include/lvtinydom.h
+++ b/crengine/include/lvtinydom.h
@@ -986,6 +986,8 @@ public:
     void getAbsRect( lvRect & rect, bool inner=false );
     /// sets node rendering structure pointer
     void clearRenderData();
+    /// reset node rendering structure pointer for sub-tree
+    void clearRenderDataRecursive();
     /// calls specified function recursively for all elements of DOM tree
     void recurseElements( void (*pFun)( ldomNode * node ) );
     /// calls specified function recursively for all elements of DOM tree matched by matchFun