--- a/crengine/src/lvmemman.cpp
+++ b/crengine/src/lvmemman.cpp
@@ -15,12 +15,17 @@
 #include "../include/lvmemman.h"
 #include "../include/lvref.h"
 #include "../include/lvtinydom.h"
+
 #ifdef _LINUX
 #ifndef _XOPEN_SOURCE
 #define _XOPEN_SOURCE
 #endif
 #include <signal.h>
 #include <unistd.h>
+#endif
+
+#ifdef _DEBUG
+#include <stdexcept>
 #endif
 
 static char file_to_remove_on_crash[2048] = "";
@@ -65,7 +70,13 @@ void crSetSignalHandler()
 /// default fatal error handler: uses exit()
 void lvDefFatalErrorHandler (int errorCode, const char * errorText )
 {
-    fprintf( stderr, "FATAL ERROR #%d: %s\n", errorCode, errorText );
+    char strbuff[10];
+    sprintf(strbuff, "%d", errorCode);
+    fprintf( stderr, "FATAL ERROR #%s: %s\n", strbuff, errorText );
+#ifdef _DEBUG
+    std::string errstr = std::string("FATAL ERROR #") + std::string(strbuff) + std::string(": ") + std::string(errorText);
+    throw std::runtime_error(errstr);
+#endif
     exit( errorCode );
 }