… cr3gui/data/hyph/Dutch.pattern
… cr3gui/data/hyph/Spanish.pattern
… crengine/include/cr3version.h
… crengine/include/crsetup.h
… crengine/include/lvtextfm.h
… crengine/src/hyphman.cpp
… crengine/src/lvdocview.cpp
… crengine/src/lvstring.cpp
… crengine/src/lvstsheet.cpp
… crengine/src/lvtextfm.cpp
… crengine/src/lvtinydom.cpp
… thirdparty/chmlib/src/chm_lib.c