Upstream fixes and cleanup from @virxkane , taken from his not-yet-merged https://github.com/virxkane/coolreader/commits/koreader-merge-post branch.
Discussion at https://github.com/koreader/crengine/pull/329#issuecomment-650019550.
See individual commit messages for details and comments.

I did not totally remove `hasCacheFile()` / `setBoxingWishedButPreventedByCache()` - I'll do that when I'm perfectly sure we won't ever need it.

`globalHash: exclude some settings on initial loading`
On a first book opening, avoid recomputing styles because text lang was set to the default of en_US, and book is FR. (This was hardly noticable on a normal book, but with a crappy book with ugly CSS that take 40 seconds to be applied, load time being reduces from 80 to 40s is noticable :)

`CSS: optimize pseudoclass nth-child() and friends`
(Hope I didn't messed up, I didn't test them all, I just tested nth-child() with the sample below.)

I have a (not really huge) book which has:
```css
 .itemtiret > p:nth-child(2) ,
 .itemtiret > p:nth-child(3) ,
 .itemtiret > p:nth-child(4) ,
 .itemtiret > p:nth-child(5) ,
 .itemtiret > p:nth-child(6) ,
 .itemtiret > p:nth-child(7) ,
 .itemtiret > p:nth-child(8) ,
 .itemtiret > p:nth-child(9) ,
 .itemtiret > p:nth-child(10) ,
 /* and the same for 6 others .itemalpha .itemrom .itemautre... */
```
And this book HTML does not even have a single node with these itemtiret, itemalpha class names!
But it was still checking all P. So, for the 282th paragraph in a DIV, it would walk all its previous 281 siblings with the non-recursive unboxing walker, the same way for these 60 selectors!

Loading time before this commit was 108s on the emulator, 200s on my Kobo.
With this commit, It goes down to 4s on the emulator and 6s on my Kobo :)

Just did a test on the emulator with the book mentionned at https://github.com/koreader/crengine/pull/276#issuecomment-517920207 (still with the 2 of the 3 huge CSS empty), which has a few such pseudoclasses, but only applied to table elements (so, a lot less than P in the previous book). Not such a noticable gain, 39s+30s => 36s + 28s.

Before:
```
06/29/20-21:57:42 DEBUG CreDocument: loading document...
CRE: document loaded, but styles re-init needed (cause: peculiar CSS pseudoclasses met)
06/29/20-21:58:21 DEBUG CreDocument: loading done.
06/29/20-21:58:21 DEBUG Typography lang: not overriding en-US with doc language: en-US
06/29/20-21:58:21 DEBUG CreDocument: rendering document...
CRE: styles re-init needed after load, re-rendering
06/29/20-21:58:51 DEBUG CreDocument: rendering done.
```

After:
```
06/29/20-22:02:20 DEBUG CreDocument: loading document...
CRE: document loaded, but styles re-init needed (cause: peculiar CSS pseudoclasses met)
06/29/20-22:02:56 DEBUG CreDocument: loading done.
06/29/20-22:02:56 DEBUG Typography lang: not overriding en-US with doc language: en-US
06/29/20-22:02:56 DEBUG CreDocument: rendering document...
CRE: styles re-init needed after load, re-rendering
06/29/20-22:03:24 DEBUG CreDocument: rendering done.
```

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/koreader/crengine/349)
<!-- Reviewable:end -->