@poire-z 2020-06-29
> (Hope I didn't messed up, I didn't test them all, I just tested nth-child() with the sample below.)

Well, my conscience hit me, and I went checking. And it might be not all good...
Possibly just because I'm not clearing what I have cached when `:last-child` or `:only-of-type` and this happens:
`CRE: document loaded, but styles re-init needed (cause: peculiar CSS pseudoclasses met)`

I'll recheck and fix it tomorrow.