@poire-z 2020-06-30
Not bumping yet.
I'd like to tackle a bit of the 109 clang static analyzer warnings...
https://travis-ci.org/github/koreader/crengine/builds/703425091
(including some real errors of mine that I didn't see because of the huge numbers :/).

@virxkane @Frenzie @NiLuJe @pkb:
any idea how to solve or silent the ~ 40 ones involving:
https://github.com/koreader/crengine/blob/ff3640d75a5016b2493b6fde1adb64a4826eb0ad/crengine/include/lvref.h#L82-L90

It's some kind of wrapper object with reference counting that is supposed to delete when no more reference. But sometimes, we use it to get the real pointer, to later get a member, so we get a new reference (to something that already exists), increase the ref count, keep the pointer we want, decrease the ref count (when the thing gets out of scope) - and it should still exists.
But the clang analyzer checks a path where the refcount decrease would reach 0 (which it never will the way we use it), and complain later when we're using the real pointer that it might have been deleted.

At least, that's what I'm undestanding.
I'm not really sure if it's just clang analyzer that check too much too dumb'ly, but I don't see a way to tell it to not checks some path or condition.
Or if it's something somewhere else, like some missing ptr=NULL after we delete, that makes clang considers this path.
Any thoughts ?
@virxkane : have you seen and considered these ones when you were solving the other sanatizer complains ?