@Frenzie 2020-06-30
@poire-z If it's a false alarm you can always add something like `// NOLINT (+clang-analyzer-unix.NewDelete)` to make it shut up about it. There are some edge cases where it gets it wrong and it looks like this might be one of them.