@virxkane 2020-06-30
@poire-z 
> any idea how to solve or silent the ~ 40 ones involving:

Sorry, I have no idea how to fix this bug. From time to time I use clang analyzer, see tons of errors and warnings, fixing one or two problems and... going to the next task.

> have you seen and considered these ones when you were solving the other sanatizer complains ?

No, I fix all bugs that address sanitizer show me, excepting memory leaks. Why you no use asan, clang also supports him?