@virxkane 2020-06-30
I think this is a real bug, but we just don't create appropriate condition for him. Statical analyzer must process ALL possible branches of the code execution, even they unreachable in real life.