@NiLuJe 2020-06-30
Clang's SA *can* trigger a wild array of false-positives, so if the codepath taken sounds a bit too wild/unrealistic, feel free to ignore it ;).

On the other hand, the runtime ASAN stuff (either Clang's or GCC's, Clang's should be a bit further ahead) should be much more detailed, and actually report stuff that actually happened ^^.
The obvious downside is that it's runtime ;).