@NiLuJe 2020-06-30
Actually running scan-build manually will allow you run spawn the web UI, showing you the exact codepath taken to arrive at that conclusion.

(No idea if tidy has a similar behavior, and/or if one is a superset of the other. i've never really used -tidy because I don't do C++, but I *haveù use the SA via scan-build).