@poire-z 2020-06-30
> Why you no use asan, clang also supports him?

Well, I'm not much interested in mastering all these debugging helps :) I'm fine with these analyze on PR, and with valgring when I get crashes and I'm really in the dark.

>  Statical analyzer must process ALL possible branches of the code execution, even they unreachable in real life.

Yes, that's what I always assumed, and as we don't get crash in real life, I'm fine with that.
I'd just like to not have these making noise :) and understand a bit why they are triggered.