@poire-z 2020-06-30
For example, we have 3 different occurences of a same exact clang path pattern. Showing the 2 I'm a bit familiar with the code:
![image](https://user-images.githubusercontent.com/24273478/86157067-20f03600-bb07-11ea-9614-abe49c9e4e9e.png)

![image](https://user-images.githubusercontent.com/24273478/86157151-44b37c00-bb07-11ea-90e9-a27cccf9fcda.png)

They both involve looping (from end to start, with a strange `--i`) a `LVPtrVector <something>` to delete the `something` - which has internal stuff delete'd too. And I'm pretty sure in both above cases that internal stuff is private and don't leak in the other something-s part of the LVPtrVector.
And I don't get why clang would complain I'm deleting released memory on a next item, after I have processed a previous item...
(and there tons of other use of LVPtrVector  in the code, that don't generate any complain).
Any idea what could clang be thinking?

I'll try setting _next, _rules, _data to NULL after the delete, but I know it's bad practice and shouldn't be needed in a destructor... (I'll have it checked on travis, I don't want to install the whole clang family on my small dev machine)