@NiLuJe 2020-06-30
That just vaguely looks like it's looping on the same index twice?