@NiLuJe 2020-06-30
If you *do* re-use the Vector contents, you should *definitely* avoid dangling pointers by assigning to nullptr after delete.

(Said the guy who deosn't do C++, so, err, take that with a grain of salt ;p).