@poire-z 2020-06-30
> That just vaguely looks like it's looping on the same index twice?

Yeah, I would expect that if I were not advancing in the indexes.
I've been suspecting the `--i` - as I would naturally use `i--`, but i guess in that context, they just do the same thing, increment `i` - and the value returned (before or after incrementation) has no meaning in the 3rd item of a for (a;b;c) loop. Right ?