@poire-z 2020-06-30
Another thing:
ftp://ftp.irisa.fr/pub/mirrors/OpenBSD/src/gnu/llvm/tools/clang/www/analyzer/faq.html#null_pointer
> The reason the analyzer often thinks that a pointer can be null is because the preceding code checked compared it against null. So if you are absolutely sure that it cannot be null, remove the preceding check 

So, if you by fear added a check for a thing to not be null, clang will assume it can be null and will take paths with it NULL in many other places, where you forgot to live in fear.
So, we should be fully living in fear, or never.

> So then, because you are lukewarm, and neither cold nor hot, I will vomit you out of My mouth
-- clang*d-tidy