@NiLuJe 2020-06-30
> Yeah, I would expect that if I were not advancing in the indexes.
> I've been suspecting the `--i` - as I would naturally use `i--`, but i guess in that context, they just do the same thing, increment `i` - and the value returned (before or after incrementation) has no meaning in the 3rd item of a for (a;b;c) loop. Right ?

Yeah, in a for loop, that's always executed at the end of an iteration, so not an issue. It'd be another thing in a while loop, though ;).