@poire-z 2020-07-01
> I'll try setting _next, _rules, _data to NULL after the delete, but I know it's bad practice and shouldn't be needed in a destructor... 

Actually, this did silence these warnings :/